Prismjs removes the code tag inside the pre tag.
I was using the prism-js plugin, I found out that the three backticks aren't converted to `<pre><code></code></pre>` but rather only to `<pre></pre>`. Firstly, <code> tag is important for HTML semantics .  Secondly, without the `<code>` tag, `<span>` elements become the direct children of `<pre>` tag and the horizontal scrolling feature provided by Prism's CSS for the `<pre>` element with `overflow-x: auto` has no effect. This is because `<span>` elements wrap instead of overflowing. This is not the intended behaviour when using Prism.js.